+++ src/core/client/views/inventory/html/app.js
@@ -148,6 +148,18 @@ const app = new Vue({
                 return;
             }
 
+            // Right-Click
+            if (e.button === 2) {
+                if (!e.target.id || e.target.id === '') {
+                    return;
+                }
+
+                if ('alt' in window) {
+                    alt.emit('inventory:Use', e.target.id, this.pageIndex);
+                }
+                return;
+            }
+
             this.dragging = true;
 
             // Calculate Element Size
@@ -368,6 +380,10 @@ const app = new Vue({
                 }
 
                 return Object.keys(target.data).map((key) => {
+                    if (key === 'event') {
+                        return { key: 'Consumeable', value: true };
+                    }
+
                     return { key, value: target.data[key] };
                 });
             }
@@ -380,6 +396,10 @@ const app = new Vue({
                 }
 
                 return Object.keys(target.data).map((key) => {
+                    if (key === 'event') {
+                        return { key: 'Consumeable', value: true };
+                    }
+
                     return { key, value: target.data[key] };
                 });
             }
@@ -392,6 +412,10 @@ const app = new Vue({
                 }
 
                 return Object.keys(target.data).map((key) => {
+                    if (key === 'event') {
+                        return { key: 'Consumeable', value: true };
+                    }
+
                     return { key, value: target.data[key] };
                 });
             }
@@ -403,6 +427,10 @@ const app = new Vue({
             }
 
             return Object.keys(target.data).map((key) => {
+                if (key === 'event') {
+                    return { key: 'Consumeable', value: true };
+                }
+
                 return { key, value: target.data[key] };
             });
         }
@@ -451,7 +479,8 @@ const app = new Vue({
                     quantity: Math.floor(Math.random() * 10),
                     weight: Math.floor(Math.random() * 5),
                     data: {
-                        water: 100
+                        water: 100,
+                        event: 'test'
                     }
                 }
             ];
@@ -503,7 +532,10 @@ const app = new Vue({
                     icon: 'hat',
                     slot: 0,
                     quantity: Math.floor(Math.random() * 10),
-                    weight: Math.floor(Math.random() * 5)
+                    weight: Math.floor(Math.random() * 5),
+                    data: {
+                        event: 'test'
+                    }
                 }
             ]);
         }