+++ src/core/shared/configurations/shared.ts
@@ -3,6 +3,8 @@ export const SHARED_CONFIG = {
     RESPAWN_TIME: 25000,
     // Blips
     MAX_BLIP_STREAM_DISTANCE: 650,
+    // PLAYERS
+    FORCE_SCREEN_SHAKE: true, // Used to shake the player's screen. Makes most aimbots a hell of a lot worse.
     // Interaction
     INTERACTION_ALWAYS_ON: false, // Determines whether or not a player needs to hit `alt` to turn on interaction.
     MAX_INTERACTION_RANGE: 8